requirements.txt
halo>=0.0.31
praw>=7.2.0
PyYAML>=5.4.1
PyGithub>=1.55
